build(generic-worker)!: support windows arm64, drop windows 386
