+++ changelog/issue-7947.md
@@ -1,0 +1,5 @@
+audience: worker-deployers
+level: major
+reference: issue 7947
+---
+Generic Worker (Windows): Windows/386 platform is now unsupported and no longer built/released. Windows arm64 support has been added as a Tier-2 platform (guaranteeing to build).
