+++ ui/docs/reference/workers/generic-worker/support-tiers.mdx
@@ -38,7 +38,7 @@ The Tier-2 platforms are:
   * generic-worker-multiuser-freebsd-amd64
   * generic-worker-multiuser-freebsd-arm64
   * generic-worker-multiuser-linux-arm64
-  * generic-worker-multiuser-windows-386
+  * generic-worker-multiuser-windows-arm64
   * generic-worker-insecure-darwin-amd64
   * generic-worker-insecure-darwin-arm64
   * generic-worker-insecure-freebsd-amd64
